traffic_accidents/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-dotenv
flask-smorest
flask-jwt-extended
passlib
marshmallow
werkzeug
